[LCR-342] refactor: bestItem 컴포넌트 분리
